chat-room/requirements.txt
fastapi
uvicorn
pymongo
python-jose
websocket-client
PyJWT
